+++ Assets/AchievementsTable.cs
@@ -1,8 +1,0 @@
-public static class AchievementsTable
-{
-    public static string ABC = "ABC";
-    public static string CDE = "CDE";
-    public static string KAKKA = "kakka";
-    public static string HAHA = "Haha";
-    public static string LOLO = "lolo";
-}