+++ Assets/Packages/com.studio23.ss2.achievementsystem/Runtime/Core/AchievementSystem.cs
@@ -9,8 +9,6 @@ namespace Studio23.SS2.AchievementSystem.Core
 
         public AchievementProvider _achievementProvider;
 
-        
-
         private void Awake()
         {
             Instance = this;